+++ lib/nmxp_crc32.c
@@ -1,7 +1,7 @@
 
 #include "nmxp_crc32.h"
 
-static unsigned long crc32_tab[] = {
+static uint32_t crc32_tab[] = {
 	0x00000000L, 0x77073096L, 0xee0e612cL, 0x990951baL, 0x076dc419L,
 	0x706af48fL, 0xe963a535L, 0x9e6495a3L, 0x0edb8832L, 0x79dcb8a4L,
 	0xe0d5e91eL, 0x97d2d988L, 0x09b64c2bL, 0x7eb17cbdL, 0xe7b82d07L,
@@ -57,10 +57,10 @@ static unsigned long crc32_tab[] = {
 };
 
 
-unsigned long crc32(const unsigned char *s, unsigned int len)
+uint32_t crc32(const unsigned char *s, unsigned int len)
 {
 	unsigned int i;
-	unsigned long crc32val;
+	uint32_t crc32val;
 
 	crc32val = 0;
 	for (i = 0;  i < len;  i ++)
